@@ -1,33 +1,39 @@
-import axios from "axios";
-import { AuthResponse } from "../store/auth/types";
+import axios from 'axios';
+import { AuthResponse } from '../store/auth/types';
 
-export const API_URL = process.env.REACT_APP_BASE_URL
+export const API_URL = process.env.REACT_APP_BASE_URL;
 
 const $api = axios.create({
-    withCredentials: true,
-    baseURL: API_URL
-})
+  withCredentials: true,
+  baseURL: API_URL,
+});
 
 $api.interceptors.request.use((config) => {
-    config.headers.Authorization = `Bearer ${localStorage.getItem('token')}`
-    return config
-})
+  config.headers.Authorization = `Bearer ${localStorage.getItem('token')}`;
+  return config;
+});
 
-$api.interceptors.response.use((config) => {
-    return config
-}, async (err) => {
+$api.interceptors.response.use(
+  (config) => {
+    return config;
+  },
+  async (err) => {
     const originalRequest = err.config;
     if (err.response.status == 401 && err.config && !err.config._isRetry) {
-        originalRequest._isRetry = true
-        try {
-            const response = await axios.get<AuthResponse>(`${process.env.REACT_APP_BASE_URL}/user/refresh`, { withCredentials: true })
-            localStorage.setItem('token', response.data.accessToken)
-            return $api.request(originalRequest)
-        } catch (error) {
-            console.log('Пользователь не авторизован', error);
-        }
-        throw err;
+      originalRequest._isRetry = true;
+      try {
+        const response = await axios.get<AuthResponse>(
+          `${process.env.REACT_APP_BASE_URL}/user/refresh`,
+          { withCredentials: true }
+        );
+        localStorage.setItem('token', response.data.accessToken);
+        return $api.request(originalRequest);
+      } catch (error) {
+        console.log('Пользователь не авторизован', error);
+      }
+      throw err;
     }
-})
+  }
+);
 
-export default $api
+export default $api;
