@@ -1,14 +1,14 @@
 declare module '*.module.css' {
-    const classes: { [key: string]: string };
-    export default classes;
+  const classes: { [key: string]: string };
+  export default classes;
 }
 
 declare module '*.module.sass' {
-    const classes: { [key: string]: string };
-    export default classes;
+  const classes: { [key: string]: string };
+  export default classes;
 }
 
 declare module '*.module.scss' {
-    const classes: { [key: string]: string };
-    export default classes;
+  const classes: { [key: string]: string };
+  export default classes;
 }
