@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import './index.css';
-import { store } from './store'
-import { Provider } from 'react-redux'
+import { store } from './store';
+import { Provider } from 'react-redux';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -12,9 +12,9 @@ const root = ReactDOM.createRoot(
 root.render(
   <>
     {/* <React.StrictMode> */}
-        <Provider store={store}>
-          <App />
-        </Provider>
+    <Provider store={store}>
+      <App />
+    </Provider>
     {/* </React.StrictMode> */}
   </>
 );
