@@ -1,10 +1,10 @@
-import { EmployeeList } from "../components/EmployeeList/EmployeeList";
+import { EmployeeList } from '../components/EmployeeList/EmployeeList';
 
 export const Employees = () => {
-    return (
-        <>
-        <h2>Employees</h2>
-        <EmployeeList/>
-        </>
-    );
-}
+  return (
+    <>
+      <h2>Employees</h2>
+      <EmployeeList />
+    </>
+  );
+};
