@@ -1,10 +1,10 @@
-import { UserList } from "../components/UserList/UserList";
+import { UserList } from '../components/UserList/UserList';
 
 export const Users = () => {
-    return (
-        <>
-        <h2>Users</h2>
-        <UserList/>
-        </>
-    );
-}
+  return (
+    <>
+      <h2>Users</h2>
+      <UserList />
+    </>
+  );
+};
