@@ -1,17 +1,17 @@
 export interface AuthResponse {
-    accessToken: string,
-    refreshToken: string,
-    user: User
+  accessToken: string;
+  refreshToken: string;
+  user: User;
 }
 
 export interface User {
-    id: string,
-    name: string,
-    email: string
+  id: string;
+  name: string;
+  email: string;
 }
 
 export interface AuthState {
-    user: object,
-    isAuth: boolean,
-    isLoading: boolean
+  user: object;
+  isAuth: boolean;
+  isLoading: boolean;
 }
