@@ -1,3 +1,3 @@
-import { RootState } from "..";
+import { RootState } from '..';
 
-export const selectEmployees = (state: RootState) => state.employees.employees
+export const selectEmployees = (state: RootState) => state.employees.employees;
