@@ -1,11 +1,11 @@
 export interface Employee {
-    _id: string;
-    email: string;
-    name: string,
+  _id: string;
+  email: string;
+  name: string;
 }
 
-export type Employees = Employee[]
+export type Employees = Employee[];
 
 export interface EmployeesState {
-    employees: Employees;
+  employees: Employees;
 }
