@@ -1,13 +1,12 @@
-import { configureStore, combineReducers } from '@reduxjs/toolkit'
-import authReducer from './auth/authSlice'
+import { configureStore, combineReducers } from '@reduxjs/toolkit';
+import authReducer from './auth/authSlice';
 import usersSlice from './users/usersSlice';
 import employeesSlice from './employees/employeesSlice';
-
 
 const rootReducer = combineReducers({
   auth: authReducer,
   users: usersSlice,
-  employees: employeesSlice
+  employees: employeesSlice,
 });
 
 export const store = configureStore({
@@ -16,4 +15,3 @@ export const store = configureStore({
 
 export type RootState = ReturnType<typeof rootReducer>;
 export type AppDispatch = typeof store.dispatch;
-
