@@ -1,3 +1,3 @@
-import { RootState } from "..";
+import { RootState } from '..';
 
-export const selectUsers = (state: RootState) => state.users.users
+export const selectUsers = (state: RootState) => state.users.users;
