@@ -1,12 +1,12 @@
 export interface User {
-    _id: string;
-    email: string;
-    name: string,
-    employees: []
+  _id: string;
+  email: string;
+  name: string;
+  employees: [];
 }
 
-export type Users = User[]
+export type Users = User[];
 
 export interface UsersState {
-    users: Users;
+  users: Users;
 }
